+++ MaterialDesignExtensionsDemo/ViewModel/FileSystemControlViewModel.cs
@@ -11,12 +11,28 @@ namespace MaterialDesignExtensionsDemo.ViewModel
         private string m_selectedAction;
         private bool m_showHiddenFilesAndDirectories;
         private bool m_showSystemFilesAndDirectories;
+        private bool m_createNewDirectoryEnabled;
 
         public override string DocumentationUrl
         {
             get
             {
                 return "https://spiegelp.github.io/MaterialDesignExtensions/#documentation/filesystemcontrols";
+            }
+        }
+
+        public bool CreateNewDirectoryEnabled
+        {
+            get
+            {
+                return m_createNewDirectoryEnabled;
+            }
+
+            set
+            {
+                m_createNewDirectoryEnabled = value;
+
+                OnPropertyChanged(nameof(CreateNewDirectoryEnabled));
             }
         }
 
@@ -71,6 +87,7 @@ namespace MaterialDesignExtensionsDemo.ViewModel
             m_selectedAction = null;
             m_showHiddenFilesAndDirectories = false;
             m_showSystemFilesAndDirectories = false;
+            m_createNewDirectoryEnabled = false;
         }
     }
 }