+++ MaterialDesignExtensionsTests/FileNameTest.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using MaterialDesignExtensions.Controllers;
+using MaterialDesignExtensions.Model;
+
+namespace MaterialDesignExtensionsTests
+{
+    [TestClass]
+    public class FileNameTest
+    {
+        [TestMethod]
+        public void TestCheckFileName()
+        {
+            Assert.AreEqual(true, FileNameHelper.CheckFileName("myDirectory"));
+            Assert.AreEqual(true, FileNameHelper.CheckFileName("myTextFile.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName(""));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName(null));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_<_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_>_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_:_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_\"_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_/_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_\\_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_|_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_?_.txt"));
+            Assert.AreEqual(false, FileNameHelper.CheckFileName("t_*_.txt"));
+        }
+    }
+}